% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\name{as.name_map}
\alias{as.name_map}
\title{Name Map Coercion}
\usage{
as.name_map(x)
}
\arguments{
\item{x}{A list with named elements the same as the slots}
}
\value{
An object of class "name_map"
}
\description{
Name Map Coercion
}
\examples{
library(smart.data)
as.name_map(list(name_map = c(A = "a", B = "z"), law = rlang::expr(eval())))
}
