% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\name{as.taxonomy}
\alias{as.taxonomy}
\title{Taxonomy Coercion}
\usage{
as.taxonomy(x)
}
\arguments{
\item{x}{A list with named elements the same as the slots}
}
\value{
An object of class "taxonomy"
}
\description{
Taxonomy Coercion
}
