% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart.data.R
\name{is.smart}
\alias{is.smart}
\title{Check for Signs of Intelligence}
\usage{
is.smart(...)
}
\arguments{
\item{...}{One or more smart.data objects}
}
\value{
A vector of logical results the length of the input.
}
\description{
\code{is.smart} checks the argument \code{i} for attributes and objects indicating the "smart.data" class
}
