% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\name{is.taxonomy}
\alias{is.taxonomy}
\title{Taxonomy Validation}
\usage{
is.taxonomy(x)
}
\arguments{
\item{x}{An object of class "taxonomy" created with \code{link{taxonomy}}}
}
\value{
A logical scalar
}
\description{
Taxonomy Validation
}
