% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\docType{class}
\name{name_map-class}
\alias{name_map-class}
\alias{name_map}
\title{Smart Naming Constructor}
\value{
An object of class "name_map"
}
\description{
\code{name_map} is an S4 class to create a naming map for use with \code{\link{smart.data}}
}
\section{Slots}{

\describe{
\item{\code{name_map}}{A named list with names as the new names of \code{$data} and the values as existing names}

\item{\code{law}}{A quoted expression that, when executed, will invoke \code{\link[data.table]{setnames}} for compatibility with \code{\link{smart.data}} method \code{$enforce.rules()}}

\item{\code{state}}{A string indicating whether the name map has been processed by \code{\link{smart.data}} method \code{$enforce.rules()}}
}}

