% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_functions.R
\name{smart.start}
\alias{smart.start}
\title{Cache Initialization and Update}
\usage{
smart.start(...)
}
\arguments{
\item{...}{See \code{\link[cachem]{cache_layered}}}
}
\description{
\code{smart.start} creates or updates the smart-cache created using \code{\link[cachem]{cache_layered}}.
}
