% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart.data.R
\name{smart.upgrade}
\alias{smart.upgrade}
\title{Upgrade a Smart Object}
\usage{
smart.upgrade(..., env = globalenv(), chatty = FALSE)
}
\arguments{
\item{...}{Names or symbols of 'smart.data' objects}

\item{env}{The environment object where the smart object(s) reside}

\item{chatty}{(logical) When \code{TRUE}, additional messages are sent to the}
}
\description{
\code{smart.upgrade} assigns a new \code{smart.data} object to replace the existing object.  The primary use case for this function is when there is a package upgrade, and down-rev objects need to be upgraded.  Objects \code{$smart.rules}, \code{private$orig.data}, \code{$cache}, and \code{$name} are preserved.
}
