% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\docType{class}
\name{taxonomy-class}
\alias{taxonomy-class}
\alias{taxonomy}
\title{Smart Taxonomy Constructor
\code{taxonomy} is an S4 class to create a taxonomy object for use with \code{\link{smart.data}}.}
\value{
An object of class "taxonomy"
}
\description{
Smart Taxonomy Constructor
\code{taxonomy} is an S4 class to create a taxonomy object for use with \code{\link{smart.data}}.
}
\section{Slots}{

\describe{
\item{\code{term}}{A symbol indicating the term to be accessed with \code{\link{smart.data}} method \code{$use()}: must be of class 'language'}

\item{\code{desc}}{A text description of the term}

\item{\code{fields}}{One or more strings indicating the fields in \code{smart.data$data} that map to the taxonomy term}

\item{\code{law}}{A logical constant only provided for compatibility with \code{\link{smart.data}} method \code{$enforce.rules()}}

\item{\code{state}}{A string indicating whether the taxonomy term has been processed by \code{\link{smart.data}} method \code{$enforce.rules()}}
}}

\examples{
identifier <- new("taxonomy", term = rlang::sym("identifier"), desc = "Identifies unique instances of a type of reference")

}
