% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructs.R
\name{as.taxonomy}
\alias{as.taxonomy}
\title{Taxonomy Coercion}
\usage{
as.taxonomy(x)
}
\arguments{
\item{x}{A list, JSON string, data.* object with named elements the same as the slots in class \code{\link{taxonomy}}}
}
\value{
Objects of class "taxonomy"
}
\description{
Taxonomy Coercion
}
\examples{
library(smart.data)
as.taxonomy(list(term = "test_term", desc = "test_description"))
}
